requirements.txt
flask==3.0.0
Pillow==10.2.0
python-magic==0.4.27
exifread==3.0.0
PyPDF2==3.0.1
opencv-python-headless==4.9.0.80
numpy==1.26.4
scikit-learn==1.4.0
joblib==1.3.2
werkzeug==3.0.1
